monad-intel/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.13.3
redis==5.1.1
httpx==0.27.2
websockets==13.1
structlog==24.4.0
orjson==3.10.10
eth-utils==5.0.0
eth-account==0.13.3
eth-abi==5.1.0
pyjwt[crypto]==2.9.0
pycryptodome==3.21.0
typer==0.12.5
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
fakeredis[lua]==2.25.1
